Make use of keycloak optional
