+++ shogun-config/src/main/java/de/terrestris/shogun/config/WebSecurityConfig.java
@@ -16,45 +16,21 @@
  */
 package de.terrestris.shogun.config;
 
-import org.keycloak.adapters.springsecurity.KeycloakConfiguration;
-import org.keycloak.adapters.springsecurity.authentication.KeycloakAuthenticationProvider;
-import org.keycloak.adapters.springsecurity.config.KeycloakWebSecurityConfigurerAdapter;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.core.authority.mapping.SimpleAuthorityMapper;
-import org.springframework.security.core.session.SessionRegistryImpl;
-import org.springframework.security.web.authentication.session.RegisterSessionAuthenticationStrategy;
-import org.springframework.security.web.authentication.session.SessionAuthenticationStrategy;
+import org.springframework.security.web.util.matcher.RequestMatcher;
 
-@KeycloakConfiguration
-public abstract class WebSecurityConfig extends KeycloakWebSecurityConfigurerAdapter {
+public interface WebSecurityConfig {
 
-    @Autowired
-    public void configureGlobal(AuthenticationManagerBuilder auth) {
-        SimpleAuthorityMapper grantedAuthorityMapper = new SimpleAuthorityMapper();
-        grantedAuthorityMapper.setPrefix("ROLE_");
+    RequestMatcher csrfRequestMatcher = httpServletRequest -> {
+        String refererHeader = httpServletRequest.getHeader("Referer");
 
-        grantedAuthorityMapper.setConvertToUpperCase(true);
+        return refererHeader != null && refererHeader.endsWith("swagger-ui/index.html");
+    };
 
-        KeycloakAuthenticationProvider keycloakAuthenticationProvider = keycloakAuthenticationProvider();
-        keycloakAuthenticationProvider.setGrantedAuthoritiesMapper(grantedAuthorityMapper);
-        auth.authenticationProvider(keycloakAuthenticationProvider);
-    }
-
-    @Bean
-    @Override
-    protected SessionAuthenticationStrategy sessionAuthenticationStrategy() {
-        return new RegisterSessionAuthenticationStrategy(new SessionRegistryImpl());
-    }
-
-    @Override
-    protected void configure(HttpSecurity http) throws Exception {
-        super.configure(http);
-
+    default void configure(HttpSecurity http) throws Exception {
         customHttpConfiguration(http);
     }
 
-    protected abstract void customHttpConfiguration(HttpSecurity http) throws Exception;
+    void customHttpConfiguration(HttpSecurity http) throws Exception;
+
 }
